move thin api services from the extension package to the thinapi package
